requirements.txt
fastapi
uvicorn[standard]
sqlalchemy==2.0.*
asyncpg
alembic
pydantic-settings
python-dotenv
passlib[bcrypt]
pyjwt
redis
celery
